Server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
faster-whisper>=1.0.0
google-genai>=1.0.0
kokoro>=0.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
